Clarify the ProhibitNoAbortFunction message
The ProhibitNoAbortFunction checks for both the abort attribute and
whether the bang is used for the function command, but the error didn't
mention this.

This is quite confusing, and it took me a while to figure out why vint
was failing on two functions in my project. This should clarify it.

Fixes #232